split into modules and migrated to the latest clap version
