+++ examples/network-topology-benchmark/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 dslab-core = { path = "../../crates/dslab-core" }
 dslab-network = { path = "../../crates/dslab-network" }
 log = "0.4"
-env_logger = "0.8.4"
+env_logger = "0.10.0"
 serde = { version = "1.0", features = ["derive"] }
 sugars = "3.0.0"
-clap = { version = "3.0.14", features = ["derive"] }
+clap = { version = "4.1.9", features = ["derive"] }
